+++ src/facade/Addons.php
@@ -6,13 +6,12 @@ namespace tsaotai\addons\facade;
 use think\Facade;
 
 /**
- * Class Addons
- * @package tsaotai\addons\facade
  * @method static void load()
  * @method static void registerRoutes()
  * @method static array getAddons()
  * @method static array scanAddons()
- * @method static bool create(string $name, array $options = [])
+ * @method static void clearCache()
+ * @method static void clearPluginCache(string $name)
  */
 class Addons extends Facade
 {