+++ src/helper.php
@@ -2,7 +2,7 @@
 declare (strict_types=1);
 
 if (!function_exists('addons')) {
-    function addons(string $name = null): mixed
+    function addons(?string $name = null): mixed
     {
         if ($name === null) {
             return app('addons');
@@ -19,19 +19,5 @@ if (!function_exists('addons_path')) {
             $addonsPath = root_path('addons');
         }
         return $addonsPath . ($path ? DIRECTORY_SEPARATOR . ltrim($path, DIRECTORY_SEPARATOR) : '');
-    }
-}
-
-if (!function_exists('addons_url')) {
-    function addons_url(string $url = ''): string
-    {
-        return url('addons/' . ltrim($url, '/'));
-    }
-}
-
-if (!function_exists('addons_view')) {
-    function addons_view(string $template = '', array $vars = []): string
-    {
-        return view($template, $vars);
     }
 }